requirements.txt
appdirs==1.4.4
filelock==3.0.12
idna==2.10
imap-tools==0.20.0
python-dotenv==0.14.0
uritools==3.0.0
urlextract==1.0.0
youtube-dl==2020.7.28
